Bump mina-sshd to 2.16.0
https://github.com/apache/mina-sshd/releases/tag/sshd-2.16.0

Remove channelGroup assignment in TransportIoService since
that is now done by NettyIoService constructor.

Remove fake implementation in CompatChannelGroup because now we have
no option to pass it down.

Since we really need to make underlying ChannelGroup#add operation
to do not store Channels in order to save memory we have to override
NettyIoService#registerChannel to do it instead.

See also:
https://github.com/apache/mina-sshd/issues/663

Change-Id: Ib6f3e29695cbe5f1661b46a88a1c77b48f962799
Signed-off-by: Ivan Hrasko <ivan.hrasko@pantheon.tech>