Reimplement TransportIoSession
If we do not want to use NettyIoService we have problems to invoke
NettyIoSession constructor in ODL's TransportIoSession -
service is a parameter for session.

The purpose of TransportIoSession is to provide us with
NettyIoSession's Adapter which is required in SSHTransportStack#
onUnderlayChannelEstablished.

This adapter is ChannelInboundHandlerAdapter implementation operating
with IOHandler which operates on IoSession.

According to SSHTransportStack docs Adapter is used to route traffic
from socket to TransportSshServer/Client.

Options:
0. Assert that we need an implementation of IoSession.
1. Let TransportIoSession be child of NettyIoSession and override
methods using NettyIoService (channelActive and channelInactive).
2. Let TransportIoSession implements IoSession (requires more work) and
reimplement NettyIoSession.Adapter as well.

Choosen option is #1.

JIRA: NETCONF-1504
Change-Id: I41490a287587a4fcacdbf6fd3f93fb6d53158829
Signed-off-by: Ivan Hrasko <ivan.hrasko@pantheon.tech>