WIP: Audit SSH algorithms
This patch passes over the registries in TransportUtils, updating their
mapping to match current registry values.

This uncovers a number of unimplemented values that got by the last
update, as well at least one bug and a couple of implementation
opportunities we are not taking.

In order to prevent this from happening, we create two coverage tests,
which enumerate the 8 enums involved in mapping and checking they are
all defined (or listed as unused).

This provides a solid baseline, guarding both sides of the conversion,
so that new enum members do not slip by unnoticed.

// FIXME: add an issue

Change-Id: Ia7f6041bdfb573366ff0de23f82338a9976a4ff6
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>