Audit and refactor TransportUtils
Split handling of the 4 algorithm groups into separate classes, as each
of them is its own world, which reduces import conflicts, keeping code
clear and concise.

As we are splitting of each class, we audit the constants involved and
document where things are coming from.

This uncovers a number of unimplemented values that got by the last
update, as well at least one bug and a couple of implementation
opportunities we are not taking.

In order to prevent this from happening, we create two coverage tests,
which enumerate the 8 enums involved in mapping and checking they are
all defined (or listed as unused).

Whiler we are at it, lets ditch ImmutableMap, as there are benefits to
having Map.ofEntries() constants.

JIRA: NETCONF-1510
Change-Id: If5b014007dd7cfa136e25370f66a822513071c1c
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>
(cherry picked from commit 8c2e4a17c57f6099310960337c1b6f3169df5d84)