Audit and refactor TransportUtils
Split handling of the 4 algorithm groups into separate classes, as each
of them is its own world, which reduces import conflicts, keeping code
clear and concise.

As we are splitting of each class, we audit the constants involved and
document where things are coming from.

This uncovers a number of unimplemented values that got by the last
update, as well at least one bug and a couple of implementation
opportunities we are not taking.

In order to prevent this from happening, we create two coverage tests,
which enumerate the 8 enums involved in mapping and checking they are
all defined (or listed as unused).

Whiler we are at it, lets ditch ImmutableMap, as there are benefits to
having Map.ofEntries() constants.

// FIXME: add an issue

Change-Id: Ia7f6041bdfb573366ff0de23f82338a9976a4ff6
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>