Add transport-crypto module
This starts the process of unifying crypto handling by introducing a
commmon CertificateSupport class.

This is inherited from the transport-ssh version of KeyUtils, which is
using BouncyCastle. The reason for that is guarded by a test, which does
not explain much :(

In any case, we improve on the previous incarnation by caching the
Provider directly, side-stepping the lookup-by-String.

JIRA: NETCONF-1506
Change-Id: I5f2b586be955a3583f7558307ecf6383631232bc
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>