Remove HttpServerCodec from pipeline
HttpServerCodec provides HttpObjectEncoder which is creating chunks
when chunked encoding is used for the response.

In order to use our own implementation we have to:
- remove it from pipeline
- add our implementation before HttpServerCodec

Other possibility is to get rid of low-level operations done in
ResponseBodyOutputStream.java, especially creating chunks. Instead
we should just provide initial response with chunked header and
pushing data as needed with termination with the last empty content.

JIRA: NETCONF-1493
Change-Id: I60b1d472b2d268ba8b2f76d87c342965f399922b
Signed-off-by: Ivan Hrasko <ivan.hrasko@pantheon.tech>