Do not format ByteBuf for chunk
While its true that chunk has to be formatted according to:
https://www.rfc-editor.org/rfc/rfc9112#name-chunked-transfer-coding
its not a responsibility of our ResponseBodyOutputStream since its
done by Netty's HttpObjectEncoder which is responsible for formatting
response headers and content.

We should just provide the content of the chunk with read-only ByteBufs
for further processing down the pipeline.

JIRA: NETCONF-1493
Change-Id: I9e8f06b456c68eac1edb20111b360c2bf0b69f3e
Signed-off-by: Ivan Hrasko <ivan.hrasko@pantheon.tech>