Abort schema setup when request is cancelled
If a request is cancelled during the schema setup process,
abort all further efforts to retrieve or build the schema context.

Once a request is cancelled, the device starts to reconnect.
In this state, it is pointless to send more get-schema requests
or attempt to build the schema context from incomplete sources.

JIRA: NETCONF-1501
Change-Id: I6e3310074d1e2ad6e279fb8d3a326e28b303cb83
Signed-off-by: Peter Suna <peter.suna@pantheon.tech>