Expose ServerRequest encoding
RFC8639's establish-subscription requires its implementation to be
cognizant of 'RPC encoding', which is wrecking our encapsulation a bit.

This patch introduces ServerRequest.requestEncoding(), which addresses
this particular need and gives the requestor the responsibility to
define it.

Carry the original Content-Type along with the request
so RFC 8639 can choose encode-json/encode-xml when <encoding> is
omitted.

JIRA: NETCONF-1488
Change-Id: I7968bf792f8abe937f7c3917bb15f8f2e45ce31a
Signed-off-by: Yaroslav Lastivka <yaroslav.lastivka@pantheon.tech>
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>