WIP: Make SSH transport configurable
Draft to put config in netconf-topology, netconf-topology seem as
better place for this config as NetconfClientConfiguration is created
here, and it is easier to pass it around.

In netconf-client it is no longer possible to change NetconfClientConfiguration
in desired way.

For start, we can change default key exchange algorithms.
The default being empty which results in no change when creating
ssh client and mina-sshd default being used.

JIRA: NETCONF-1502
Change-Id: Ic5208602cccb7618026e63ca37abc1e05e08702a
Signed-off-by: Samuel Schneider <samuel.schneider@pantheon.tech>