Rework TransportIoService and TransportIoSession
We do not need a full implementations of neither
TransportIoService nor TransportIoSession.

NettyIoService's channelGroup is used to track all open channels,
so that we can close them when we are shutdown.

We do not need to do that, as we track the underlying transport
instead, hence any channels tracked here will disappear when we shut
down the underlay.

Rework TransportIoService to be just an intermediate step to create
sessions. Use simple counter for creating session ids.

Rework TransportIoSession to be an partial implementation of IoSession,
adapt NettyIoSession.Adapter implementation to provide suitable inbound
handler.

JIRA: NETCONF-1504
Change-Id: I026779eb3c2be7234b50bf94405fa071d8a446ba
Signed-off-by: Ivan Hrasko <ivan.hrasko@pantheon.tech>