Use IndentedXML to emit indentation
This ditches the use of stax-utils in favor of a YANGTOOLS-provided
utility. The new indenter does not emit a final newline in documents
so a few tests are updated to account for this -- which makes things
consistent with what JSON does.

JIRA: NETCONF-1528
Change-Id: I65550a077d4a07d0105aad7feaf70fe5de29d5b6
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>
(cherry picked from commit 03de5126a06bbbd0933bba8403c6f75fb9b4aa01)