Use QName-based encoding in restconf-server-spi
There is a large number of mapping and remapping, to the point where we
keep both the QName and MonitoringEncoding side by side.

MonitoringEncoding is left for request URL parsing, but we turn it into
a QName before talking to the backend.

This requires a minor change in the way streams are exported: if we have
a QName, which does not have a well-defined MonitoringEncoding, we just
do not report it.

JIRA: NETCONF-1487
Change-Id: Idd0c59137d8a23e1dfdbb0fed998038d5fffe20a
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>
(cherry picked from commit 8545e8b4b0687fe7f00ce9e1da0d63edb9b28843)