WIP: Add KeyPairParser
This adopts transport-tls's ConfigUtils.extractKeyPair() as
KeyPairParser.parseKeyPair().

An immediate improvement here is that we do not rely on algorithm
string, but retain the full KeyFactory, eliminating the need for
re-lookup and error paths coming from there.

There corner-case of not having a public key is marked up for future
improvement, as is the step of byte[] in case of OpenSSH public key,
as well as the need to check that the public key actually corresponds
to the private key.

JIRA: NETCONF-1506
Change-Id: Ib29a77524a5e64395d7f089de590b353bc1d9224
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>