Remove FIXME from KeepaliveSalFacade
The previous patch added an explanation of what happens if invokeRpc()
throws.

If the invoked RPC completes successfully, the keep-alive messages
are re-enabled. If the RPC throws an exception, the device will be
reconnected. The reconnection logic is handled by the RpcsDecorator

Hence, there is no need for re-enabling keep-alive messages,
as this KeepaliveSalFacade instance will be destroyed and a new one will
be created upon the next successful connection.

JIRA: NETCONF-1501
Change-Id: I2cbc2e02b3b3c423dafb0d4f7303d7c02d0e84cf
Signed-off-by: Peter Suna <peter.suna@pantheon.tech>