Merge SchemaSourceException methods in SchemaSetup
Merge handleSchemaResolutionException and
handleMissingSchemaSourceException into a single method, as they
perform the same logic. The only difference was that
addUnresolvedCapabilities was conditionally called only if
the collection was not empty. This check is unnecessary,
as addUnresolvedCapabilities does nothing if the provided collection
is empty.

SchemaSourceException is implemented only by
MissingSchemaSourceException and SchemaResolutionException,
so catching SchemaSourceException simplifies the entire process.

JIRA: NETCONF-1501
Change-Id: I3dab4436c5262bb1f820d17785aedb3f095c6b01
Signed-off-by: Peter Suna <peter.suna@pantheon.tech>